feat: separate @lodestar/utils to node and browser
