+++ packages/utils/src/bytes/node.ts
@@ -2,18 +2,6 @@ import {toBufferLE, toBigIntLE, toBufferBE, toBigIntBE} from "bigint-buffer";
 
 type Endianness = "le" | "be";
 
-const hexByByte: string[] = [];
-export function toHexString(bytes: Uint8Array): string {
-  let hex = "0x";
-  for (const byte of bytes) {
-    if (!hexByByte[byte]) {
-      hexByByte[byte] = byte < 16 ? "0" + byte.toString(16) : byte.toString(16);
-    }
-    hex += hexByByte[byte];
-  }
-  return hex;
-}
-
 /**
  * Return a byte array from a number or BigInt
  */
